requirements.txt
# requirements.txt
Flask
Flask-Login
Flask-Mail
mysql-connector-python
gunicorn  # Production WSGI server